Fixes donation controller spec after merging
